+++ backend/canisters/local_user_index/api/src/lifecycle/init.rs
@@ -18,7 +18,6 @@ pub struct Args {
     pub cycles_dispenser_canister_id: CanisterId,
     pub escrow_canister_id: CanisterId,
     pub event_relay_canister_id: CanisterId,
-    pub internet_identity_canister_id: CanisterId,
     pub video_call_operators: Vec<Principal>,
     pub oc_secret_key_der: Option<Vec<u8>>,
     pub test_mode: bool,